+++ CHANGELOG.md
@@ -10,7 +10,21 @@ N/A
 
 * TDB
 
-## v0.0.10, 2019-04-03
+## v0.0.11, 2019-11-08
+
+### Breaking changes
+N/A
+
+### Migration Steps
+N/A
+
+### Changes
+
+* supports Slack channel override for every message
+* changes auto-scaling messages to Note instead of Alert on Slack. 
+* sets up prettier and eslint
+
+## v0.0.10, 2019-11-07
 
 ### Breaking changes
 N/A